pyproject.toml
[project]
name = "confluence-kb"
version = "0.1.0"
description = "Data pipeline to extract and prepare knowledge base from Confluence"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-flink>=2.0.0",
    "atlassian-python-api>=4.0.3",
    "confluent-kafka>=2.10.0",
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "markitdown[all]>=0.1.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "selectolax>=0.3.16",
    "parsel>=1.8.1",
    "sentence-transformers>=4.1.0",
    "typer>=0.15.2",
    "pydantic-settings>=2.9.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
]
